Implementation of the balloon option for SpatialLines and SpatialPolygons.
Also modified the way this functionality is launched. It now tests that
there is data in the object, and does not display long/lat anymore. Next
step would be to display the actual attributes that have been called for
display (eg if log(zinc) has been called, print the log(zinc) value and
not the zinc value).



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/plotkml@92 edb9625f-4e0d-4859-8d74-9fd3b1da38cb